Fix NPE in iterator, fix empi-rules.json, add extra tests
