+++ hapi-fhir-jpaserver-empi/src/test/java/ca/uhn/fhir/jpa/empi/provider/EmpiProviderMergePersonsR4Test.java
@@ -5,8 +5,6 @@ import ca.uhn.fhir.empi.api.EmpiMatchResultEnum;
 import ca.uhn.fhir.empi.util.AssuranceLevelUtil;
 import ca.uhn.fhir.rest.server.exceptions.InvalidRequestException;
 import ca.uhn.fhir.rest.server.exceptions.ResourceNotFoundException;
-import org.hl7.fhir.r4.model.Bundle;
-import org.hl7.fhir.r4.model.Patient;
 import org.hl7.fhir.r4.model.Person;
 import org.hl7.fhir.r4.model.StringType;
 import org.junit.jupiter.api.BeforeEach;
@@ -37,18 +35,6 @@ public class EmpiProviderMergePersonsR4Test extends BaseProviderR4Test {
 		myFromPersonId = new StringType(myFromPerson.getIdElement().getValue());
 		myToPerson = createPerson();
 		myToPersonId = new StringType(myToPerson.getIdElement().getValue());
-	}
-
-	@Test
-	public void testMatch() {
-		Patient jane = buildJanePatient();
-		jane.setActive(true);
-		Patient createdJane = createPatient(jane);
-		Patient newJane = buildJanePatient();
-
-		Bundle result = myEmpiProviderR4.match(newJane);
-		assertEquals(1, result.getEntry().size());
-		assertEquals(createdJane.getId(), result.getEntryFirstRep().getResource().getId());
 	}
 
 	@Test
