+++ hapi-fhir-jpaserver-empi/src/test/resources/empi/empi-rules.json
@@ -1,7 +1,18 @@
 {
 	"version": "1",
 	"candidateSearchParams": [
-
+		{
+			"resourceType": "Patient",
+			"searchParams": ["birthdate"]
+		},
+		{
+			"resourceType": "*",
+			"searchParams": ["identifier"]
+		},
+		{
+			"resourceType": "Patient",
+			"searchParams": ["general-practitioner"]
+		}
 	],
 	"candidateFilterSearchParams": [
 		{
